+++ tests/genomics_tests/test_genomic_pipeline.py
@@ -61,7 +61,7 @@ from rdr_service.model.questionnaire_response import QuestionnaireResponse, Ques
 from rdr_service.genomic.genomic_job_controller import GenomicJobController
 from rdr_service.genomic.genomic_job_components import GenomicFileIngester
 from rdr_service.genomic.genomic_mappings import array_file_types_attributes
-from rdr_service.offline import genomic_pipeline
+from rdr_service.offline.genomics import genomic_pipeline, genomic_dispatch
 from rdr_service.participant_enums import (
     SampleStatus,
     Race,
@@ -570,7 +570,7 @@ class GenomicPipelineTest(BaseTestCase):
         ])
 
         # Execute from cloud task
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         # metrics insert called via cloud task
         self.assertEqual(metric_cloud_task.call_count, 2)
@@ -642,7 +642,7 @@ class GenomicPipelineTest(BaseTestCase):
         task_data['file_data']['file_path'] = f"{bucket_name}/{subfolder}/{updated_aw2_filename}"
 
         # Simulate new file uploaded
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         # NEW metrics insert called via cloud task
         self.assertEqual(metric_cloud_task.call_count, 4)
@@ -763,8 +763,8 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Execute from cloud task
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data_seq)
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data_gen)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data_seq)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data_gen)
 
         # metrics insert called via cloud task
         self.assertEqual(metric_cloud_task.call_count, 4)
@@ -814,8 +814,8 @@ class GenomicPipelineTest(BaseTestCase):
                 self.member_dao.update(member)
 
         # Execute from cloud task
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data_aw5_wgs)
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data_aw5_array)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data_aw5_wgs)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data_aw5_array)
 
         # Test the fields against the DB
         gc_metrics = self.metrics_dao.get_all()
@@ -1066,7 +1066,7 @@ class GenomicPipelineTest(BaseTestCase):
         self._create_stored_samples([(2, 1002)])
 
         # Execute from cloud task
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # run_id = 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # run_id = 1 & 2
 
         # metrics insert called via cloud task
         self.assertEqual(metric_cloud_task.call_count, 1)
@@ -1144,7 +1144,7 @@ class GenomicPipelineTest(BaseTestCase):
         self._create_stored_samples([(2, 1002)])
 
         # Execute from cloud task
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # run_id = 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # run_id = 1 & 2
 
         # metrics insert called via cloud task
         self.assertEqual(metric_cloud_task.call_count, 1)
@@ -1235,7 +1235,7 @@ class GenomicPipelineTest(BaseTestCase):
             pipelineId=config.GENOMIC_DEPRECATED_WGS_DRAGEN
         )
         # Execute from cloud task
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # run_id = 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # run_id = 1 & 2
 
         # metrics insert called via cloud task
         self.assertEqual(metric_cloud_task.call_count, 1)
@@ -2119,7 +2119,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         # Test the data was ingested OK
         for member in self.member_dao.get_all():
@@ -2189,7 +2189,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         # Test the data was ingested OK
         members = [m for m in self.member_dao.get_all() if m.id in [4, 5]]
@@ -2254,7 +2254,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         members = self.member_dao.get_members_with_non_null_sample_ids()
 
@@ -2296,7 +2296,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         # Test member was created
         new_member = self.member_dao.get_member_from_collection_tube(1, 'aou_wgs')
@@ -2308,10 +2308,10 @@ class GenomicPipelineTest(BaseTestCase):
         task_data['file_data']['file_path'] = f"{bucket_name}/{file_name}"
 
         # Call pipeline function twice
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 3 & 4
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 3 & 4
 
         # No new GenomicSetMembers should be inserted in this run
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 5 & 6
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 5 & 6
 
         members = self.member_dao.get_all()
         members.sort(key=lambda x: x.id)
@@ -2358,7 +2358,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         # Setup Test AW1F file
         gc_manifest_file = open_genomic_set_file("Genomic-AW1F-Workflow-Test-1.csv")
@@ -2384,7 +2384,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Ingest AW1F
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 3 & 4
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 3 & 4
 
         # Test db updated
         members = sorted(self.member_dao.get_all(), key=lambda x: x.id)
@@ -4652,7 +4652,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         # Test AW4 manifest updated fields
         for member in self.member_dao.get_all():
@@ -4685,7 +4685,7 @@ class GenomicPipelineTest(BaseTestCase):
         self.assertEqual(file_name, file_record.fileName)
 
         # Test AW4 Raw table
-        genomic_pipeline.load_awn_manifest_into_raw_table(f"{bucket_name}/{sub_folder}/{file_name}", "aw4")
+        genomic_dispatch.load_awn_manifest_into_raw_table(f"{bucket_name}/{sub_folder}/{file_name}", "aw4")
 
         aw4_dao = GenomicAW4RawDao()
         raw_records = aw4_dao.get_all()
@@ -4764,7 +4764,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         # Test AW4 manifest updated fields
         for member in self.member_dao.get_all():
@@ -4801,7 +4801,7 @@ class GenomicPipelineTest(BaseTestCase):
         self.assertEqual(file_name, file_record.fileName)
 
         # Test AW4 Raw table
-        genomic_pipeline.load_awn_manifest_into_raw_table(f"{bucket_name}/{sub_folder}/{file_name}", "aw4")
+        genomic_dispatch.load_awn_manifest_into_raw_table(f"{bucket_name}/{sub_folder}/{file_name}", "aw4")
 
         aw4_dao = GenomicDefaultBaseDao(
             model_type=GenomicAW4Raw
@@ -4874,7 +4874,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         # Test Files Processed
         file_record = self.file_processed_dao.get(1)
@@ -4956,7 +4956,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
         manifest_record = self.manifest_file_dao.get(1)
         feedback_record = self.manifest_feedback_dao.get(1)
 
@@ -5011,7 +5011,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         # Set up AW2 File
         aw2_bucket_name = _FAKE_GENOMIC_CENTER_BUCKET_A
@@ -5210,7 +5210,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_run_id 5 & 6
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_run_id 5 & 6
 
         # NEW metrics insert called via cloud task
         self.assertEqual(cloud_task.call_count, 5)
@@ -5364,7 +5364,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         gc_manifest_file = open_genomic_set_file("Genomic-GC-Manifest-Workflow-Test-6.csv")
 
@@ -5384,7 +5384,7 @@ class GenomicPipelineTest(BaseTestCase):
                 folder=_FAKE_GENOTYPING_FOLDER,
             )
 
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         all_incidents = self.incident_dao.get_all()
         correct_incident = list(filter(lambda x: gc_manifest_filename in x.message, all_incidents))
@@ -5410,7 +5410,7 @@ class GenomicPipelineTest(BaseTestCase):
 
         test_file_path = f"{_FAKE_GENOMIC_CENTER_BUCKET_A}/{_FAKE_GENOTYPING_FOLDER}/{aw1_manifest_filename}"
         # Run load job
-        genomic_pipeline.load_awn_manifest_into_raw_table(test_file_path, "aw1")
+        genomic_dispatch.load_awn_manifest_into_raw_table(test_file_path, "aw1")
 
         # Expected columns in table
         expected_columns = [
@@ -5479,7 +5479,7 @@ class GenomicPipelineTest(BaseTestCase):
                 folder=_FAKE_GENOTYPING_FOLDER,
             )
             test_file_path = f"{_FAKE_GENOMIC_CENTER_BUCKET_A}/{_FAKE_GENOTYPING_FOLDER}/{aw1_manifest_filename}"
-            genomic_pipeline.load_awn_manifest_into_raw_table(test_file_path, "aw1")
+            genomic_dispatch.load_awn_manifest_into_raw_table(test_file_path, "aw1")
             time.sleep(5)
 
         biobank_id = '2'
@@ -5526,7 +5526,7 @@ class GenomicPipelineTest(BaseTestCase):
         test_file_path = f"{_FAKE_GENOMIC_CENTER_BUCKET_A}/{_FAKE_BUCKET_FOLDER}/{test_file_name}"
 
         # Run load job
-        genomic_pipeline.load_awn_manifest_into_raw_table(test_file_path, "aw2")
+        genomic_dispatch.load_awn_manifest_into_raw_table(test_file_path, "aw2")
 
         aw2_raw_records = self.aw2_raw_dao.get_all()
 
@@ -5584,7 +5584,7 @@ class GenomicPipelineTest(BaseTestCase):
         test_file_path = f"{_FAKE_GENOMIC_CENTER_BUCKET_A}/{_FAKE_BUCKET_FOLDER}/{test_file_name}"
 
         # Run load job
-        genomic_pipeline.load_awn_manifest_into_raw_table(test_file_path, "aw2")
+        genomic_dispatch.load_awn_manifest_into_raw_table(test_file_path, "aw2")
 
         aw2_raw_records = self.aw2_raw_dao.get_all()
 
@@ -5655,7 +5655,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         incident_dao = GenomicIncidentDao()
 
@@ -5704,7 +5704,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         incident_dao = GenomicIncidentDao()
         incidents = incident_dao.get_all()
@@ -5737,7 +5737,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Execute from cloud task
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         incident_dao = GenomicIncidentDao()
         incidents = incident_dao.get_all()
@@ -5804,7 +5804,7 @@ class GenomicPipelineTest(BaseTestCase):
         ])
 
         # Execute from cloud task
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         current_run = list(filter(lambda x: x.jobId == job_id, self.job_run_dao.get_all()))[0]
         self.assertEqual(current_run.runResult, GenomicSubProcessResult.ERROR)
@@ -5851,7 +5851,7 @@ class GenomicPipelineTest(BaseTestCase):
         self._create_stored_samples([(2, 1002)])
 
         # Execute from cloud task
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # run_id = 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # run_id = 1 & 2
         current_run = list(filter(lambda x: x.jobId == GenomicJob.METRICS_INGESTION, self.job_run_dao.get_all()))[0]
         self.assertEqual(current_run.runResult, GenomicSubProcessResult.ERROR)
 
@@ -5887,7 +5887,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         # Test the data was ingested OK
         files_processed = self.file_processed_dao.get_all()
@@ -5909,7 +5909,7 @@ class GenomicPipelineTest(BaseTestCase):
             folder=_FAKE_GENOTYPING_FOLDER,
         )
         test_file_path = f"{_FAKE_GENOMIC_CENTER_BUCKET_A}/{_FAKE_GENOTYPING_FOLDER}/{aw1_manifest_filename}"
-        genomic_pipeline.load_awn_manifest_into_raw_table(test_file_path, "aw1")
+        genomic_dispatch.load_awn_manifest_into_raw_table(test_file_path, "aw1")
 
         raw_records = self.aw1_raw_dao.get_all()
         raw_records = [obj for obj in raw_records if obj.sample_id != ""]
@@ -6248,7 +6248,7 @@ class GenomicPipelineTest(BaseTestCase):
         )
 
         # Run load job
-        genomic_pipeline.load_awn_manifest_into_raw_table(test_file_path, "aw2")
+        genomic_dispatch.load_awn_manifest_into_raw_table(test_file_path, "aw2")
 
         self.data_generator.create_database_genomic_file_processed(
             runId=1,
@@ -6258,7 +6258,7 @@ class GenomicPipelineTest(BaseTestCase):
             fileName=test_file_name,
         )
 
-        genomic_pipeline.load_awn_manifest_into_raw_table(test_file_path, "aw2")
+        genomic_dispatch.load_awn_manifest_into_raw_table(test_file_path, "aw2")
 
         genomic_pipeline.reconcile_raw_to_aw2_ingested()
 
@@ -6294,10 +6294,10 @@ class GenomicPipelineTest(BaseTestCase):
             }
         }
         # Run load job
-        genomic_pipeline.load_awn_manifest_into_raw_table(f"{_FAKE_GENOMIC_CENTER_BUCKET_A}/{file_name}", "aw1")
+        genomic_dispatch.load_awn_manifest_into_raw_table(f"{_FAKE_GENOMIC_CENTER_BUCKET_A}/{file_name}", "aw1")
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)  # job_id 1 & 2
 
         records = self.aw1_raw_dao.get_all()
         self.assertEqual(0, len(records))
@@ -6388,7 +6388,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data)
 
         # Setup AW2 investigation file
         aw2_manifest_file = open_genomic_set_file("RDR_AoU_GEN_TestDataManifestInvestigation.csv")
@@ -6424,7 +6424,7 @@ class GenomicPipelineTest(BaseTestCase):
         }
 
         # Call pipeline function
-        genomic_pipeline.execute_genomic_manifest_file_pipeline(task_data_aw2)
+        genomic_dispatch.execute_genomic_manifest_file_pipeline(task_data_aw2)
 
         # metrics insert called via cloud task
         self.assertEqual(metric_cloud_task.call_count, 2)
